+++ program/table_body_builder.gd
@@ -1,4 +1,4 @@
-# body_builder.gd
+# table_body_builder.gd
 # This file is part of I, Voyager
 # https://ivoyager.dev
 # *****************************************************************************
@@ -17,7 +17,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # *****************************************************************************
-class_name IVBodyBuilder
+class_name IVTableBodyBuilder
 extends RefCounted
 
 ## Builds [IVBody] instances from data tables.
@@ -115,9 +115,8 @@ var flag_fields := {
 }
 
 
-var BodyScript: Script
 
-var _orbit_builder: IVOrbitBuilder
+var _orbit_builder: IVTableOrbitBuilder
 var _composition_builder: IVCompositionBuilder
 var _table_name: StringName
 var _row: int
@@ -125,16 +124,13 @@ var _real_precisions := {}
 
 
 func _ivcore_init() -> void:
-	BodyScript = IVGlobal.procedural_classes[&"Body"]
-	_orbit_builder = IVGlobal.program[&"OrbitBuilder"]
+	_orbit_builder = IVGlobal.program[&"TableOrbitBuilder"]
 	_composition_builder = IVGlobal.program.get(&"CompositionBuilder")
 
 
-func build_from_table(table_name: String, row: int, parent: IVBody) -> IVBody: # Main thread!
+func build_body_from_table(body: IVBody, table_name: String, row: int, parent: IVBody) -> void:
 	_table_name = table_name
 	_row = row
-	@warning_ignore("unsafe_method_access")
-	var body: IVBody = BodyScript.new()
 	body.name = IVTableData.get_db_entity_name(table_name, row)
 	_set_flags_from_table(body, parent)
 	_set_orbit_from_table(body, parent)
@@ -144,7 +140,6 @@ func build_from_table(table_name: String, row: int, parent: IVBody) -> IVBody: #
 	if enable_precisions:
 		body.characteristics[&"real_precisions"] = _real_precisions
 		_real_precisions = {} # reset for next body
-	return body
 
 
 func _set_flags_from_table(body: IVBody, parent: IVBody) -> void: