+++ CHANGELOG.md
@@ -6,6 +6,20 @@ File format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 See cloning and downloading instructions [here](https://www.ivoyager.dev/developers/).
 
+## [v0.0.19] - Not Released
+
+Under development using Godot 4.2.2.
+
+Requires plugin [ivoyager_table_reader](https://github.com/ivoyager/ivoyager_table_importer) v0.0.7.
+
+### Added
+* API support for adding IVSmallBodiesGroup data by code (not just tables/binaries).
+
+### Changed
+* [API breaking] Improved builder class names with "source" and "what": "BinaryAsteroidsBuilder", "TableBodyBuilder", "TableOrbitBuilder", etc.
+* Tables specified in IVCoreSettings.body_tables no longer need to be top-down ordered.
+
+
 ## [v0.0.18] - 2024-03-15
 
 Developed using Godot 4.2.1. _Has backward breaking changes!_
@@ -59,4 +73,5 @@ Requires plugin [ivoyager_table_reader](https://github.com/ivoyager/ivoyager_tab
 ##
 I, Voyager projects v0.0.16 and earlier used a different core submodule [ivoyager](https://github.com/ivoyager/ivoyager) (now depreciated); see previous changelog [here](https://github.com/ivoyager/ivoyager/blob/master/CHANGELOG.md).
 
+[v0.0.19]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.18...HEAD
 [v0.0.18]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.17...v0.0.18